API

Modules

NestedNumbers.NestedNumbers

Types and constants

NestedNumbers.ABool

NestedNumbers.AComplex

NestedNumbers.AInteger

NestedNumbers.AReal

NestedNumbers.BareNumber

NestedNumbers.NestedNumber

NestedNumbers.OpaqueNumber

NestedNumbers.QBool

NestedNumbers.QComplex

NestedNumbers.QInteger

NestedNumbers.QReal

NestedNumbers.QuantityNumber

NestedNumbers.RemoteNumber

Functions and macros

Documentation

NestedNumbers.NestedNumbers — Module

NestedNumbers

Numbers that kind of contain numbers.

NestedNumbers.BareNumber — Type

abstract type BareNumber{T<:Number} end

Super-type for numbers do not have a unit or similar "decoration".

NestedNumbers.NestedNumber — Type

abstract type NestedNumber{T<:Number} end

Super-type for numbers that contain another number of type T, directly or semantically.

The value of the inner number may or may not be directly accessible, and may or may not have a unit or similar, depending of the subtype of NestedNumber.

Examples include opaque numbers, quantities with units, etc.

NestedNumbers.OpaqueNumber — Type

abstract type OpaqueNumber{T<:Number} end

Super-type for numbers that do not have an accessible value, for example traced or symbolic numbers.

NestedNumbers.QuantityNumber — Type

abstract type QuantityNumber{T<:Number} <: NestedNumber{T} end

Read as "maybe static Integer", an `Integ Super-type for number-like quantities (e.g. numbers with units).

NestedNumbers.RemoteNumber — Type

abstract type RemoteNumber{T<:Number} end

Super-type for numbers that are physically stored on another system or device, so that accessing the value may take time.

NestedNumbers.ABool — Type

const ABool = Union{Bool, BareNumber{Bool}}

Read as "almost Bool", a Bool or NestedNumber{Bool}.

NestedNumbers.AComplex — Type

const AComplex{T<:Real} = Union{Complex{T}, BareNumber{Complex{T}}}

Read as "almost Complex", a Complex or BareNumber{<:Complex}.

NestedNumbers.AInteger — Type

const AInteger{T<:Integer} = Union{T, BareNumber{T}}

Read as "almost Integer", an Integer or NestedNumber{<:Integer}.

NestedNumbers.AReal — Type

const AReal{T<:Real} = Union{T, BareNumber{T}}

Real as "almost Real", a Real or NestedNumber{<:Real}.

NestedNumbers.QBool — Type

const QBool = Union{Bool, BareNumber{Bool}, QuantityNumber{Bool}, QuantityNumber{<:BareNumber{Bool}}}

Read as "boolean or boolean-valued quantity", a Bool or QuantityNumber{Bool}.

NestedNumbers.QComplex — Type

const QComplex{T<:Real} = Union{Complex{T}, BareNumber{Complex{T}}, QuantityNumber{Complex{T}}, QuantityNumber{<:BareNumber{Complex{T}}}}

Read as "complex or complex-valued quantity", a Complex or QuantityNumber{<:Complex}.

NestedNumbers.QInteger — Type

const QInteger{T<:Integer} = Union{T, BareNumber{T}, QuantityNumber{T}, QuantityNumber{<:BareNumber{T}}}

Read as "integer or integer-valued quantity", an Integer or QuantityNumber{<:Integer}.

NestedNumbers.QReal — Type

const QReal{T<:Real} = Union{T, BareNumber{T}, QuantityNumber{T}, QuantityNumber{<:BareNumber{T}}}

Read as "real number or real-valued quantity", a Real or QuantityNumber{<:Real}.